refactor: razor cuts - remove experimental bard and flatten tools
- Removed `src/experimental/bard.rs` and the `Bard` feature (code-to-story translation).
- Removed `src/tools/cache.rs` and `src/tools/ui.rs` and merged them into `src/tools/runner.rs` as private implementation details.
- Updated `src/lib.rs`, `src/tools/cli.rs`, and `src/main.rs` to reflect the removals.
- Updated `.jules/razor.md` with reduction metrics.
- Verified with full test suite.
