+++ src/lib.rs
@@ -90,6 +90,4 @@ pub mod semantic;
 pub mod text;
 pub mod tools;
 
-pub mod experimental;
-
 pub use tools::highlight;
