+++ src/main.rs
@@ -7,7 +7,7 @@ use miette::Result;
 
 use glossa::tools::cli::{Cli, Commands};
 use glossa::tools::repl::run_repl;
-use glossa::tools::runner::{bard_file, build_file, check_file, highlight_file, run_file};
+use glossa::tools::runner::{build_file, check_file, highlight_file, run_file};
 
 fn main() -> Result<()> {
     let cli = Cli::parse();
@@ -32,10 +32,6 @@ fn main() -> Result<()> {
 
         Some(Commands::Highlight { input }) => {
             highlight_file(&input)?;
-        }
-
-        Some(Commands::Bard { input }) => {
-            bard_file(&input)?;
         }
 
         Some(Commands::Repl) | None => {
