+++ src/tools/mod.rs
@@ -1,6 +1,4 @@
-pub mod cache;
 pub mod cli;
 pub mod highlight;
 pub mod repl;
 pub mod runner;
-pub mod ui;
